# openapi.yaml
openapi: 3.1.0
info:
  title: Rawae Al-Haram API
  description: |
    Complete API specification for the Rawae Al-Haram water donation and delivery management system.

    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## User Roles
    - **admin**: Full system access
    - **donor**: Can make donations and view donation history
    - **mosque_admin**: Manages mosque needs and tank images
    - **auditor**: Can verify donations and access reports
    - **investor**: Access to impact metrics and funnel data
    - **logistics_supervisor**: Manages trucks and deliveries
  version: 1.0.0
  contact:
    name: API Support
    email: support@rawae-alharam.com
servers:
  - url: http://localhost:8000/api
    description: API Base URL
...
paths:
...
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            example:
              email: user@example.com
      responses:
        '200':
          description: Password reset link sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                success: true
                message: If the email exists, a password reset link has been sent.
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password with token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            example:
              token: reset-token
              email: user@example.com
              password: newpassword123
              password_confirmation: newpassword123
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                success: true
                message: Password has been reset successfully.
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: Invalid or expired reset token.
        '422':
          $ref: '#/components/responses/ValidationError'

  # Mosques Routes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication. Token includes user_id, role, email claims.
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            message: Validation error
            errors:
              field_name:
                - The field name is required.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
      required:
        - success
        - message
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
      required:
        - success
        - message
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required:
        - success
        - message
        - errors
...
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        password_confirmation:
          type: string
          format: password
      required:
        - token
        - email
        - password
        - password_confirmation

    # User Schemas
...